+++ core/CHANGELOG.md
@@ -1,4 +1,5 @@
 # TBD
+* remove freeHostPortTracker and all host-container port mappings
 * Make tests declare a timeout and mark them as failed if they don't complete in that time
 
 # 0.3.1